requirements.txt
pandas
numpy
statsmodels
boto3
python-dotenv
